Clarify the purpose of default answers in -testing
